# [Proof of Concept] `Polyester.@batch` for threading