Have you considered using Polyester.jl for threading instead of `Threads`?
I always use the code in this PR when I want to do fair benchmarks of my implementations against CellListMap.jl. I get consistently better performance with this PR over `main`.

Here is a very cheap benchmark that is running one time step of an n-body simulation. This is the code:
https://github.com/trixi-framework/PointNeighbors.jl/blob/main/benchmarks/n_body.jl#L15-L40

This is on an AMD Ryzen Threadripper 3990X using 128 threads:
<img width="746" alt="grafik" src="https://github.com/m3g/CellListMap.jl/assets/44124897/ecc7a6a8-5e25-47e4-aa85-b4bedae22d25">
This is a minimum speedup of 1.6x and a speedup of 4.5x for the largest problem size.

Here is a real-life benchmark, running one particle interaction loop of the Weakly Compressible SPH method:
<img width="747" alt="grafik" src="https://github.com/m3g/CellListMap.jl/assets/44124897/68479506-c482-4efd-82e1-109de61c83da">
This is a speedup of 2x for most problem sizes.