@efaulhaber 2024-07-06
> Concerning the benchmark: is that using CellListMap? That was not clear to me.

Yes, `foreach_point_neighbor` in the code I linked above just forwards to `map_pairwise!` in this case.