@lmiq 2024-07-06
Now that I think, probably what `@batch` is doing there is adding `@inbounds`, `@fastmath`. I don't think that performance difference is associated to the threading per se.