@lmiq 2024-07-13
Offtopic: can you please send me the link of your talk at JuliaCon about this when it is available? Thanks.