@efaulhaber 2024-07-15
Back to the topic:
I was confused as well, since `Polyester.@batch` is actually slower in our code for large problems.
So I repeated the benchmark with `Threads.@threads` and `Threads.@threads :static` and found this:
<img width="679" alt="grafik" src="https://github.com/user-attachments/assets/7f19a00c-2171-44b0-8084-28b826bab505">
So `Threads.@threads :static` is just as fast as `Polyester.@batch`, except for very small problems, exactly as I would expect.
Interestingly, slapping `@threads :static` to the serial loop is significantly faster than the `nbatches` approach in `main`. This is a 2x speedup for 1M particles and a 4.7x speedup for 65M particles in the cheap n-body benchmark.

But as you already noticed, the difference becomes smaller with more neighbors. In this example, I have a search radius of 3x the particle spacing, which gives me ~100 neighbors with ~3 particles per cell. When I increase the search radius to 10x the particle spacing, the speedup for 1M particles goes down from 2x to 1.11x.