@efaulhaber 2024-07-15
This is not using the `map_pairwise!` taking a `CellListPair`, right? I only changed this definition.