@efaulhaber 2024-07-16
Your benchmark on a Threadripper 3990X:
`map_pairwise_parallel!`:
```julia
julia> @btime map_pairwise($((x,y,i,j,d2,out) -> begin out[i] += inv(d2); return out end), $sys);
  8.930 ms (286 allocations: 54.41 KiB)
```
`map_pairwise_serial!` with `@batch`:
```julia
julia> @btime map_pairwise($((x,y,i,j,d2,out) -> begin out[i] += inv(d2); return out end), $sys);
  6.020 ms (53 allocations: 8.59 KiB)
```
`map_pairwise_serial!` with `Threads.@threads :static`:
```julia
julia> @btime map_pairwise($((x,y,i,j,d2,out) -> begin out[i] += inv(d2); return out end), $sys);
  3.438 ms (698 allocations: 132.58 KiB)
```